chatbot-ai-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
openai-agents==0.10.2
pydantic-settings==2.9.1
python-dotenv==1.0.0
loguru==0.7.3
pytest>=8.2,<9
pytest-asyncio>=0.25
